Update dllmain.cpp
Capitalized the signatures, no real reason to do so, just me being stupid :)
Also added a check to make sure that the file "f" was actually created before we try to close it. Also no real point to this, just wanted to get rid of the warning in Visual Studio.